+++ hardwarelibrary/powermeters/__init__.py
@@ -1,3 +1,4 @@
 
 from .powermeterdevice import PowerMeterDevice
 from .integradevice import IntegraDevice
+from .fieldmasterdevice import FieldMasterDevice, DebugFieldMasterDevice